index/pyproject.toml
[tool.poetry]
name = "index"
version = "0.1.0"
description = ""
authors = ["Ng Guoyou <gyng@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
reverse-geocode = "^1.4.1"
exifread = "^3.0.0"
click = "^8.1.7"
ultralytics = "^8.0.184"
torch = "^2.1.1"
fast-colorthief = "^0.0.5"
p-tqdm = "^1.4.0"
transformers = "^4.44.2"
torchvision = "^0.19.1"
einops = "^0.8.0"
safetensors = "^0.4.5"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
